system: BReactor: don't store a callback void pointer in BSmallTimer, instead just pass the BSmallTimer pointer to the callback.
To preserver memory efficiency, switch from SAvl to CAvl and reorder members. To preserve time efficiency, keep a flag indicating 
whether a timer is really a BTimer or not, and directly call the BTimer_handler if it is (as oppossed to having an internal 
callback which calls the BTimer_handler). Additionally, store the timer state in a new member instead of in absTime, as the 
new layout allows so without increasing the size of BSmallTimer.